Remove the parser.
Reader generates always the same thing: a form.
Inside the form we have the underlying data/value,
so every form need to be evaluated.
With this, we can remove the parser step.

The caveat (for now) is that I introduced
a bunch of "if is_a?(Form)", but I can handle those later.
THEY HAVE TO GO! But we can carry on developing the language :D