Context is around, env is inside.
Decisions, decisions... I think is worth to semantically separate
context and environment here.

If the binding is not available in the environment,
it may be available in the surrounding context.
The surroundings can be brought to you by a closure
or a normal invocation.

If context was created via closure
them this context itself has it's own context... and so on.